Break apart test_utils.js file and place functions elsewhere.
