+++ tests/04_compile_test.js
@@ -1,4 +1,4 @@
-let test_utils = require('./test_utils');
+let user_utils = require('./user_utils');
 let model_building_utils = require('./model_building_utils');
 let user = "./users/testuser/";
 
@@ -11,18 +11,18 @@ module.exports = {
     },
 
     'Login' : function (client) {
-        test_utils.login(client);
+        user_utils.login(client);
     },
 
     'Compile AS' : function (client) {
         let filename = 'Formalisms/__LanguageSyntax__/SimpleClassDiagram/SimpleClassDiagramMM.model';
-        test_utils.load_model(client, [filename]);
+        model_building_utils.load_multiple_models(client, [filename]);
         model_building_utils.compile_model(client, "AS", folder_name, "classDiagram.metamodel");
     },
 
     'Compile CS' : function (client) {
         let filename = 'Formalisms/__LanguageSyntax__/SimpleClassDiagram/SimpleClassDiagram.umlIcons.model';
-        test_utils.load_model(client, [filename]);
+        model_building_utils.load_multiple_models(client, [filename]);
         model_building_utils.compile_model(client, "CS", folder_name, "classDiagram.umlIcons.metamodel");
     },
